+++ src/components/pages/index.js
@@ -1,2 +1,0 @@
-export { default } from "./pages.component";
-export { default as Page } from "./page/page.component";